fix(golem): honest agent-memory notice when session unavailable (#257)
