+++ cmd/golem/memory_test.go
@@ -3,7 +3,9 @@ package main
 import (
 	"bytes"
 	"context"
+	"database/sql"
 	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
@@ -14,6 +16,26 @@ import (
 	"github.com/kstruzzieri/go-llm/memory"
 	"github.com/kstruzzieri/go-llm/provider"
 )
+
+// openMemoryStore prepares the hardened DB file, opens it WAL-mode single-conn,
+// runs migrations, and re-secures the file. Test helper: production opens the
+// shared handle via openMemoryRuntime instead.
+func openMemoryStore(ctx context.Context, dbPath string) (*memory.SQLiteStore, *sql.DB, error) {
+	db, err := openMemoryDB(ctx, dbPath)
+	if err != nil {
+		return nil, nil, err
+	}
+	store, err := memory.NewStore(ctx, db)
+	if err != nil {
+		_ = db.Close()
+		return nil, nil, fmt.Errorf("golem: init memory store: %w", err)
+	}
+	if err := chmodDBFiles(dbPath); err != nil {
+		_ = db.Close()
+		return nil, nil, err
+	}
+	return store, db, nil
+}
 
 func TestResultMessagesRedactsMemorySearch(t *testing.T) {
 	const secret = "USER PREFERS A SECRET THING"
